processing module is renamed to custom
