+++ kaggle_sets/bin_classification.py
@@ -1,5 +1,5 @@
 from kaggle_sets.model_service.classifier_service import ClassifierService
-from kaggle_sets.processing.functions.metrics import ConfusionMatrix
+from kaggle_sets.custom.functions.metrics import ConfusionMatrix
 
 
 def train_save_classifier():
