+++ kaggle_sets/custom/__init__.py
@@ -1,0 +1,1 @@
+# TODO: rename "custom" module to "customs"
