+++ kaggle_sets/custom/models/lin_regressor.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-from kaggle_sets.processing.functions.loss_functions import LossEnum, LossFunction
+from kaggle_sets.custom.functions.loss_functions import LossEnum, LossFunction
 from kaggle_sets.exceptions.exceptions import ModelParameterError
-from kaggle_sets.processing.models.model import Model
-import kaggle_sets.processing.preprocessing.datasplits as ds
-from kaggle_sets.processing.models.optimizers import SGD, Optimizer
-import kaggle_sets.processing.preprocessing.data_scalar as scal
+from kaggle_sets.custom.models.model import Model
+import kaggle_sets.custom.preprocessing.datasplits as ds
+from kaggle_sets.custom.models.optimizers import SGD, Optimizer
+import kaggle_sets.custom.preprocessing.data_scalar as scal
 import json
