--- kaggle_sets/processing/__init__.py
@@ -1,1 +1,0 @@
-# TODO: rename "processing" module to "customs"
